Add test for issue #4
This test exposes the failure described in issue #4. It might turn out
to be too silent for comfortable debugging, but we can add logging and
test set-up helpers later as needed.

Of course, @benwr is much more familiar with the code than me, so he
might be able to find out what is going on without much digging.