Implement choose_context_to_advance_promise
This does the job, I think, but it's ugly. I will see if I can tidy it
up.

Also, it's still not totally lazy:

Suppose you have a promise P and only one pending context C, which can
fulfill P. When you pass P to choose_context_to_advance_promise, it will
always choose C, even when no promisee is registered. This is necessary
for the root promise, because no promisee is ever registered on it.

This small bit of non-laziness should be fine, though, because hey, why
wouldn't you choose a context that can fulfill exactly the promise that
you want fulfilled?

I considered adding a promisee for the root promise, but for this I
would have to change Reply or add a special root super-context. Both
would be more awkward than the current solution. (Although, thinking
about it now, the super-context might be not as bad.)