laziness: Fix workspace unlocking and non-root promises
Remove is_blocking, because it was redundant in "self.is_blocking and
any(promisee_context can_advance_promise)".

Make unlocking of workspaces work. I had forgotten about that in the
previous version.

Most importantly, the previous version didn't handle cases where the
promise to be advanced could be advanced by a context that is blocked by
a context that is not its descendant. Okay, nobody understands this.

        X–B
       /
    TBA
       \
        P

TBA is the context to be advanced. P wants to unlock B's answer. The
previous version wouldn't recognize B as blocking P, because P isn't B's
ancestor. This commit should fix that, because now the recursion doesn't
follow the ancestor relation, but the promisee relation.

The graph isn't completely accurate, but I hope it illustrates my point.