REFACTOR can_advance_promise: Combine promises
Andreas suggested to combine the workspace's answer_promise and
final_workspace_promise in a common construct. Do this and merge
can_fulfill_promise into can_advance_promise, because it would be too
trivial.

I thought of adding `Workspace.get_promisees(db)` in order to get rid of
one of the `for`s in the list comprehension. But so far Workspace has a
small signature and none of its methods takes a Datastore as an
argument. So I think it's better not to add one.