EXPERIMENT Remove successor
It annoys me to track successor and other_contexts separately. (In the
context of reflection.) So I tried to get rid of the successor by
putting everything in one list. This works for easy cases, but causes
trouble with the Automation. Put it on ice for now, because I don't want
to take the time to understand and make it work.